Complete native parity and memory safety gates

+++ native/standalone/src/mellowrt_debug.c
@@ -6,11 +6,11 @@ int mellowrt_debug_log_snapshot(void *user, const MDebugSnapshot *snapshot) {
     (void)user;
     if (!snapshot || !snapshot->insn) return 1;
     fprintf(stderr,
-            "[mellowrt] pc=%u opcode=%u stack=%zu locals=%zu frames=%zu\n",
+            "[mellowrt] pc=%u opcode=%u stack=%llu locals=%llu frames=%llu\n",
             snapshot->pc,
             (unsigned)snapshot->insn->opcode,
-            snapshot->stack_len,
-            snapshot->locals_len,
-            snapshot->frame_len);
+            (unsigned long long)snapshot->stack_len,
+            (unsigned long long)snapshot->locals_len,
+            (unsigned long long)snapshot->frame_len);
     return 1;
 }
